# 003 Code/Aiary-Backend-Public-main/swagger-api.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/openai-test/diary": {
      "post": {
        "tags": [
          "OpenAI Test"
        ],
        "summary": "일기 생성 테스트",
        "description": "메시지 컨텍스트를 기반으로 OpenAI를 사용하여 일기를 생성합니다",
        "operationId": "generateDiary",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiaryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "일기 생성 실패",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "200": {
            "description": "일기 생성 성공",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/openai-test/stream": {
      "get": {
        "tags": [
          "OpenAI Test"
        ],
        "summary": "스트리밍 채팅 테스트",
        "description": "OpenAI를 사용하여 실시간 스트리밍 채팅 응답을 반환합니다",
        "operationId": "streamChat",
        "parameters": [
          {
            "name": "message",
            "in": "query",
            "description": "사용자 메시지",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "스트리밍 시작 성공",
            "content": {
              "text/event-stream": {}
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DiaryRequest": {
        "type": "object",
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageDto"
            }
          }
        }
      },
      "MessageDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string"
          },
          "content": {
            "type": "string"
          }
        }
      },
...
    }
  }
}